misc/HtmlTestSuite/results/test-0497.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>CSS Grid Layout Test: Grid items with percentage margins</title>
<link href="mailto:rego@igalia.com" rel="author" title="Manuel Rego Casasnovas"/>
<link href="https://drafts.csswg.org/css-grid-1/#item-margins" rel="help"/>
<meta content="Checks grid items percentage margins are resolved against the inline size of their grid area (in a track with fixed max sizing function and intrinsic min)." name="assert"/>
<link href="/css/support/grid.css" rel="stylesheet"/>
<link href="/fonts/ahem.css" rel="stylesheet" type="text/css"/>
<style>
.grid {
  font: 10px/1 Ahem;
  grid-template-columns: minmax(auto, 100px);
  width: 500px;
  justify-items: start;
  position: relative;
}

.grid &gt; div:nth-child(1) { background: cyan; }
.grid &gt; div:nth-child(2) {
  background: magenta;
  width: 100%;
  height: 10px;
}

.marginLeft50Percent { margin-left: 50%; }
.marginRight50Percent { margin-right: 50%; }
.marginTop50Percent { margin-top: 50%; }
.marginBottom50Percent { margin-bottom: 50%; }
</style>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>
<script type="text/javascript">
  setup({ explicit_done: true });
</script>

</head><body onload="document.fonts.ready.then(() =&gt; { checkLayout('.grid'); })">

<div id="log"/>

<h3>Direction LTR</h3>

<pre>Item margin-left: 50%;</pre>

<div class="grid">
  <div class="marginLeft50Percent" data-expected-height="10" data-expected-margin-left="50" data-expected-width="10" data-offset-x="50">X</div>
  <div data-expected-height="10" data-expected-width="100" data-offset-x="0" data-offset-y="10"/>
</div>

<pre>Item margin-right: 50%;</pre>

<div class="grid">
  <div class="marginRight50Percent" data-expected-height="10" data-expected-margin-right="50" data-expected-width="10" data-offset-x="0">X</div>
  <div data-expected-height="10" data-expected-width="100" data-offset-x="0" data-offset-y="10"/>
</div>

<pre>Item margin-top: 50%;</pre>

<div class="grid">
  <div class="marginTop50Percent" data-expected-height="10" data-expected-margin-top="50" data-expected-width="10" data-offset-y="50">X</div>
  <div data-expected-height="10" data-expected-width="100" data-offset-x="0" data-offset-y="60"/>
</div>

<pre>Item margin-bottom: 50%;</pre>

<div class="grid">
  <div class="marginBottom50Percent" data-expected-height="10" data-expected-margin-bottom="50" data-expected-width="10" data-offset-y="0">X</div>
  <div data-expected-height="10" data-expected-width="100" data-offset-x="0" data-offset-y="60"/>
</div>

<h3>Direction RTL</h3>

<pre>Item margin-left: 50%;</pre>

<div class="grid directionRTL">
  <div class="marginLeft50Percent" data-expected-height="10" data-expected-margin-left="50" data-expected-width="10" data-offset-x="490">X</div>
  <div data-expected-height="10" data-expected-width="100" data-offset-x="400" data-offset-y="10"/>
</div>

<pre>Item margin-right: 50%;</pre>

<div class="grid directionRTL">
  <div class="marginRight50Percent" data-expected-height="10" data-expected-margin-right="50" data-expected-width="10" data-offset-x="440">X</div>
  <div data-expected-height="10" data-expected-width="100" data-offset-x="400" data-offset-y="10"/>
</div>

<pre>Item margin-top: 50%;</pre>

<div class="grid directionRTL">
  <div class="marginTop50Percent" data-expected-height="10" data-expected-margin-top="50" data-expected-width="10" data-offset-y="50">X</div>
  <div data-expected-height="10" data-expected-width="100" data-offset-x="400" data-offset-y="60"/>
</div>

<pre>Item margin-bottom: 50%;</pre>

<div class="grid directionRTL">
  <div class="marginBottom50Percent" data-expected-height="10" data-expected-margin-bottom="50" data-expected-width="10" data-offset-y="0">X</div>
  <div data-expected-height="10" data-expected-width="100" data-offset-x="400" data-offset-y="60"/>
</div>
</body></html>
